AI-OCR-Table-Extraction/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
pandas==2.1.3
paddleocr==2.7.0
pytesseract==0.3.10
easyocr==1.7.1
torch==2.1.0
torchvision==0.16.0
pymongo==4.6.0
motor==3.3.1
python-multipart==0.0.6
ultralytics==8.0.202
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
passlib[bcrypt]==1.7.4
websockets==12.0
transformers==4.35.0
prometheus-client==0.19.0
aiofiles==23.2.1
pillow==10.1.0
